done some refactoring
split the routes into multiple files
and changed some routes to make it restfull
