+++ server.py
@@ -1,5 +1,7 @@
 from fastapi import FastAPI, Request
 from routes import pokemon_routes
+from routes import evolve_routes
+from routes import trainer_routes
 import data.models as models
 from data.database import engine
 from utils.migration import load_db
@@ -12,6 +14,9 @@ def migration():
     load_db()
 
 app.include_router(pokemon_routes.router)
+app.include_router(trainer_routes.router)
+app.include_router(evolve_routes.router)
+
 @app.middleware("http")
 async def log_req(request:Request, call_next):
     print(f'got req. to: {request.url}, method: {request.method}')
